Fix for arithmetics
