+++ csharp-basics/exercises/Arithmetic/Exercise10/Program.cs
@@ -1,38 +1,7 @@
-﻿namespace CalculateArea
+﻿using System;
+
+namespace CalculateArea
 {
-    class Geometry
-    {
-        public static double CalculateCircleArea(double radius)
-        {
-            if (radius < 0)
-            {
-                throw new ArgumentException("Error: Radius cannot be negative.");
-            }
-
-            return Math.PI * radius * radius;
-        }
-
-        public static double CalculateRectangleArea(double length, double width)
-        {
-            if (length < 0 || width < 0)
-            {
-                throw new ArgumentException("Error: Length and width cannot be negative.");
-            }
-
-            return length * width;
-        }
-
-        public static double CalculateTriangleArea(double baseLength, double height)
-        {
-            if (baseLength < 0 || height < 0)
-            {
-                throw new ArgumentException("Error: Base length and height cannot be negative.");
-            }
-
-            return 0.5 * baseLength * height;
-        }
-    }
-
     class Program
     {
         static void Main()
@@ -42,60 +11,21 @@
             int choice;
             do
             {
-                Console.WriteLine("1. Calculate the Area of a Circle");
-                Console.WriteLine("2. Calculate the Area of a Rectangle");
-                Console.WriteLine("3. Calculate the Area of a Triangle");
-                Console.WriteLine("4. Quit");
-
-                Console.Write("Enter your choice (1-4): ");
+                ShowMenu();
                 choice = int.Parse(Console.ReadLine());
 
                 switch (choice)
                 {
                     case 1:
-                        Console.Write("Enter the radius of the circle: ");
-                        double circleRadius = double.Parse(Console.ReadLine());
-                        try
-                        {
-                            double circleArea = Geometry.CalculateCircleArea(circleRadius);
-                            Console.WriteLine($"The area of the circle is: {circleArea:F2}");
-                        }
-                        catch (ArgumentException ex)
-                        {
-                            Console.WriteLine(ex.Message);
-                        }
+                        CalculateCircleArea();
                         break;
 
                     case 2:
-                        Console.Write("Enter the length of the rectangle: ");
-                        double rectangleLength = double.Parse(Console.ReadLine());
-                        Console.Write("Enter the width of the rectangle: ");
-                        double rectangleWidth = double.Parse(Console.ReadLine());
-                        try
-                        {
-                            double rectangleArea = Geometry.CalculateRectangleArea(rectangleLength, rectangleWidth);
-                            Console.WriteLine($"The area of the rectangle is: {rectangleArea:F2}");
-                        }
-                        catch (ArgumentException ex)
-                        {
-                            Console.WriteLine(ex.Message);
-                        }
+                        CalculateRectangleArea();
                         break;
 
                     case 3:
-                        Console.Write("Enter the base length of the triangle: ");
-                        double triangleBase = double.Parse(Console.ReadLine());
-                        Console.Write("Enter the height of the triangle: ");
-                        double triangleHeight = double.Parse(Console.ReadLine());
-                        try
-                        {
-                            double triangleArea = Geometry.CalculateTriangleArea(triangleBase, triangleHeight);
-                            Console.WriteLine($"The area of the triangle is: {triangleArea:F2}");
-                        }
-                        catch (ArgumentException ex)
-                        {
-                            Console.WriteLine(ex.Message);
-                        }
+                        CalculateTriangleArea();
                         break;
 
                     case 4:
@@ -109,6 +39,70 @@
 
                 Console.WriteLine();
             } while (choice != 4);
+        }
+
+        static void ShowMenu()
+        {
+            Console.WriteLine("1. Calculate the Area of a Circle");
+            Console.WriteLine("2. Calculate the Area of a Rectangle");
+            Console.WriteLine("3. Calculate the Area of a Triangle");
+            Console.WriteLine("4. Quit");
+
+            Console.Write("Enter your choice (1-4): ");
+        }
+
+        static void CalculateCircleArea()
+        {
+            Console.Write("Enter the radius of the circle: ");
+            double circleRadius = double.Parse(Console.ReadLine());
+
+            try
+            {
+                double circleArea = Geometry.CalculateCircleArea(circleRadius);
+                Console.WriteLine($"The area of the circle is: {circleArea:F2}");
+            }
+            catch (ArgumentException ex)
+            {
+                Console.WriteLine(ex.Message);
+            }
+        }
+
+        static void CalculateRectangleArea()
+        {
+            Console.Write("Enter the length of the rectangle: ");
+            double rectangleLength = double.Parse(Console.ReadLine());
+
+            Console.Write("Enter the width of the rectangle: ");
+            double rectangleWidth = double.Parse(Console.ReadLine());
+
+            try
+            {
+                double rectangleArea = Geometry.CalculateRectangleArea(rectangleLength, rectangleWidth);
+                Console.WriteLine($"The area of the rectangle is: {rectangleArea:F2}");
+            }
+            catch (ArgumentException ex)
+            {
+                Console.WriteLine(ex.Message);
+            }
+        }
+
+        static void CalculateTriangleArea()
+        {
+            Console.Write("Enter the base length of the triangle: ");
+            double triangleBase = double.Parse(Console.ReadLine());
+
+            Console.Write("Enter the height of the triangle: ");
+            double triangleHeight = double.Parse(Console.ReadLine());
+
+            try
+            {
+                double triangleArea = Geometry.CalculateTriangleArea(triangleBase, triangleHeight);
+                Console.WriteLine($"The area of the triangle is: {triangleArea:F2}");
+            }
+            catch (ArgumentException ex)
+            {
+                Console.WriteLine(ex.Message);
+            }
         }
     }
 }
