+++ csharp-basics/exercises/Arithmetic/Exercise3/Program.cs
@@ -1,4 +1,4 @@
-﻿namespace SumAverageRunningInt
+﻿namespace NumberRangeCalculator
 {
     class Program
     {
