+++ csharp-basics/exercises/Arithmetic/Exercise4/Program.cs
@@ -17,8 +17,6 @@
             }
 
             return product;
-
         }
-
     }
 }
