+++ csharp-basics/exercises/Arithmetic/Exercise5/Program.cs
@@ -1,4 +1,6 @@
-﻿namespace CalculateArea
+﻿using System;
+
+namespace NumberGuessingGame
 {
     class Program
     {
